Add local storage
